requirements.txt
amqp==5.3.1
asgiref==3.8.1
better-profanity==0.7.0
billiard==4.2.2
bleach==6.2.0
botocore==1.38.24
celery==5.5.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap4==2024.10
cron_descriptor==2.0.6
Django==5.1.13
django-allauth==65.2.0
django-appconf==1.0.6
django-celery-beat==2.8.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-htmx==1.21.0
django-recaptcha==4.0.0
django-select2==8.4.1
django-storages==1.14.4
django-timezone-field==7.1
djangorestframework==3.15.2
emoji==2.15.0
gunicorn==23.0.0
idna==3.10
inflect==7.5.0
jmespath==1.0.1
kombu==5.5.4
Markdown==3.7
more-itertools==10.6.0
oauthlib==3.2.2
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
randomcolor==0.4.4.6
redis==6.4.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typeguard==4.4.2
typing_extensions==4.12.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
